separating functions better. idk what to call it so i called it rin
